+++ luaxp.lua
@@ -22,8 +22,10 @@ _M.binops = {
     , { op='%',  prec= 3 }
     , { op='+',  prec= 4 }
     , { op='-',  prec= 4 }
-    , { op='<',  prec= 6 }
+    , { op='<<', prec= 5 }
+    , { op='>>', prec= 5 }
     , { op='..', prec= 5 }
+    , { op='<',  prec= 6 }
     , { op='<=', prec= 6 }
     , { op='>',  prec= 6 }
     , { op='>=', prec= 6 }
@@ -675,7 +677,7 @@ end
 if not bit then
     -- Adapted from "BitUtils", Lua-users wiki at http://lua-users.org/wiki/BitUtils; thank you kind stranger(s)...
     bit = {}
-    bit['nand'] = function(x,y,z)
+    bit.nand = function(x,y,z)
         z=z or 2^16
         if z<2 then
             return 1-x*y
@@ -683,10 +685,10 @@ if not bit then
             return bit.nand((x-x%z)/z,(y-y%z)/z,math.sqrt(z))*z+bit.nand(x%z,y%z,math.sqrt(z))
         end
     end
-    bit["bnot"]=function(y,z) return bit.nand(bit.nand(0,0,z),y,z) end
-    bit["band"]=function(x,y,z) return bit.nand(bit["bnot"](0,z),bit.nand(x,y,z),z) end
-    bit["bor"]=function(x,y,z) return bit.nand(bit["bnot"](x,z),bit["bnot"](y,z),z) end
-    bit["bxor"]=function(x,y,z) return bit["band"](bit.nand(x,y,z),bit["bor"](x,y,z),z) end
+    bit.bnot = function(y,z) return bit.nand(bit.nand(0, 0, z), y, z) end
+    bit.band = function(x,y,z) return bit.nand(bit.bnot(0, z), bit.nand(x, y, z), z) end
+    bit.bor = function(x,y,z) return bit.nand(bit.bnot(x, z), bit.bnot(y, z), z) end
+    bit.bxor = function(x,y,z) return bit.band(bit.nand(x, y, z), bit.bor(x, y, z), z) end
 end
 
 -- Let's get to work
@@ -989,7 +991,7 @@ local function scan_table( expr, index )
         elseif ch == ':' or ch == "=" then
             if not val then comperror("Missing key", index) end
             -- For now, only allow constants for key. May allow other expressions (e.g. variable references) later.
-            if not ( val.__type == CONST ) then comperror("Key must be constant (integer or string literal)", index) end
+            if not isAtom( val, CONST ) then comperror("Key must be constant (integer or string literal)", index) end
             key = val
             val = nil
             index = skip_white( expr, index + 1 )
@@ -1034,7 +1036,7 @@ local function scan_binop( expr, index )
                 -- matches something
                 matched = true
                 prec = f.prec
-                break;
+                break
             end
         end
         if not matched then
@@ -1419,6 +1421,12 @@ _run = function( atom, ctx, stack )
             else
                 v = bit.bxor( coerce(v1, "number"), coerce(v2, "number") )
             end
+        elseif e.op == '<<' or e.op == '>>' then
+            local f = coerce( v2, "number" )
+            if f < 0 then evalerror( "Invalid shift count "..tostring(f)) end
+            v = ( e.op == '<<' ) and
+                ( coerce( v1, "number" ) * ( 2 ^ f ) ) or
+                math.floor( coerce( v1, "number" ) / ( 2 ^ f ) )
         elseif e.op == '<' then
             if not check_operand(v1, {"number","string"}, v2) then evalerror("Invalid comparison ("
                 .. base.type(v1) .. e.op .. base.type(v2) .. ")", e.pos) end
@@ -1581,7 +1589,7 @@ _run = function( atom, ctx, stack )
             end
         elseif e.name == "push" then
             if #e.args < 2 then evalerror("push() at least two arguments are required ", e.pos) end
-            if e.args[1].__type ~= "vref" then evalerror("push() argument 1 must be array name", e.pos) end
+            if not isAtom( e.args[1], VREF ) then evalerror("push() argument 1 must be array name", e.pos) end
             local nv = e.args[1].name
             local xv = runfetch( e.args[2], ctx, stack )
             local nmax = false
@@ -1589,13 +1597,12 @@ _run = function( atom, ctx, stack )
                 nmax = runfetch( e.args[3], ctx, stack )
                 if base.type(nmax) ~= "number" or nmax < 1 then evalerror("push() argument 3 must be numeric > 0", e.pos) end
             end
-            if (ctx.__lvars or {})[nv] then
-                if base.type(ctx.__lvars[nv]) ~= "table" then evalerror("argument 1 is not array", e.pos) end
+            v = NULLATOM
+            if (ctx.__lvars or {})[nv] and not isNull( ctx.__lvars[nv] ) then
+                if base.type(ctx.__lvars[nv]) ~= "table" or isAtom( ctx.__lvars[nv] ) then evalerror("argument 1 is not array", e.pos) end
                 v = ctx.__lvars[nv]
-            elseif ctx[nv] then
-                if base.type(ctx[nv]) ~= "table" then evalerror("argument 1 is not array", e.pos) end
-                v = ctx[nv]
-            else
+            end
+            if is_non( v ) then
                 ctx.__lvars = ctx.__lvars or {}
                 v = {}
                 ctx.__lvars[nv] = v
@@ -1607,16 +1614,14 @@ _run = function( atom, ctx, stack )
             end
         elseif e.name == "pop" then
             if #e.args < 1 then evalerror("pop() missing required argument ", e.pos) end
-            if e.args[1].__type ~= "vref" then evalerror("pop() argument 1 must be array name", e.pos) end
+            if not isAtom( e.args[1], VREF ) then evalerror("pop() argument 1 must be array name", e.pos) end
             local nv = e.args[1].name
             local av
-            if (ctx.__lvars or {})[nv] then
-                if base.type(ctx.__lvars[nv]) ~= "table" then evalerror("argument 1 is not array", e.pos) end
+            if (ctx.__lvars or {})[nv] and not isNull( ctx.__lvars[nv] ) then
+                if base.type(ctx.__lvars[nv]) ~= "table" or isAtom( ctx.__lvars[nv] ) then evalerror("argument 1 is not array", e.pos) end
                 av = ctx.__lvars[nv]
-            elseif ctx[nv] then
-                if base.type(ctx[nv]) ~= "table" then evalerror("argument 1 is not array", e.pos) end
-                av = ctx[nv]
-            else
+            end
+            if is_non( av ) then
                 ctx.__lvars = ctx.__lvars or {}
                 av = {}
                 ctx.__lvars[nv] = av
@@ -1624,7 +1629,7 @@ _run = function( atom, ctx, stack )
             v = table.remove( av ) or NULLATOM
         elseif e.name == "unshift" then
             if #e.args < 2 then evalerror("unshift() at least two arguments are required ", e.pos) end
-            if e.args[1].__type ~= "vref" then evalerror("unshift() argument 1 must be array name", e.pos) end
+            if not isAtom( e.args[1], VREF ) then evalerror("unshift() argument 1 must be array name", e.pos) end
             local nv = e.args[1].name
             local xv = runfetch( e.args[2], ctx, stack )
             local nmax = false
@@ -1632,13 +1637,12 @@ _run = function( atom, ctx, stack )
                 nmax = runfetch( e.args[3], ctx, stack )
                 if base.type(nmax) ~= "number" or nmax < 1 then evalerror("unshift() argument 3 must be numeric > 0", e.pos) end
             end
-            if (ctx.__lvars or {})[nv] then
-                if base.type(ctx.__lvars[nv]) ~= "table" then evalerror("argument 1 is not array", e.pos) end
+            v = NULLATOM
+            if (ctx.__lvars or {})[nv] and not isNull( ctx.__lvars[nv] ) then
+                if base.type(ctx.__lvars[nv]) ~= "table" or isAtom( ctx.__lvars[nv] ) then evalerror("argument 1 is not array", e.pos) end
                 v = ctx.__lvars[nv]
-            elseif ctx[nv] then
-                if base.type(ctx[nv]) ~= "table" then evalerror("argument 1 is not array", e.pos) end
-                v = ctx[nv]
-            else
+            end
+            if is_non( v ) then
                 ctx.__lvars = ctx.__lvars or {}
                 v = {}
                 ctx.__lvars[nv] = v
@@ -1650,16 +1654,14 @@ _run = function( atom, ctx, stack )
             end
         elseif e.name == "shift" then
             if #e.args < 1 then evalerror("shift() missing required argument", e.pos) end
-            if e.args[1].__type ~= "vref" then evalerror("shift() argument 1 must be array name", e.pos) end
+            if not isAtom( e.args[1], VREF ) then evalerror("shift() argument 1 must be array name", e.pos) end
             local nv = e.args[1].name
             local av
-            if (ctx.__lvars or {})[nv] then
-                if base.type(ctx.__lvars[nv]) ~= "table" then evalerror("argument 1 is not array", e.pos) end
+            if (ctx.__lvars or {})[nv] and not isNull( ctx.__lvars[nv] ) then
+                if base.type(ctx.__lvars[nv]) ~= "table" or isAtom( ctx.__lvars[nv] ) then evalerror("argument 1 is not array", e.pos) end
                 av = ctx.__lvars[nv]
-            elseif ctx[nv] then
-                if base.type(ctx[nv]) ~= "table" then evalerror("argument 1 is not array", e.pos) end
-                av = ctx[nv]
-            else
+            end
+            if is_non( av ) then
                 ctx.__lvars = ctx.__lvars or {}
                 av = {}
                 ctx.__lvars[nv] = av